jatom.js: fix 'link' population
Fix jatom.js to adhere to RFC 4287 for 'link' population.

Before this change, the logic for finding the 'link' for
an atom entry would just grab the first 'link'-type node
from a given entry.  But an entry may have multiple 'link'
nodes; one needs to grab the '<link rel="alternate" ....'
node for that case.

An example feed that uses multiple 'link' nodes (emitted
'link' field is wrong before this change):

 * http://blog.racket-lang.org/feeds/posts/default